#375: Moved the PassiveEquipmentStatsIncrease Parser to the model folder
The file was not modified yet.
